core: Slight code reorganisation
* Update the license preambles
* Reorder the includes to match the c/c++ rules of cpplint
* Removes useless includes
* Removes unused functions
