+++ core/oio_core.h
@@ -1,7 +1,6 @@
 /*
 OpenIO SDS core library
-Copyright (C) 2014 Worldine, original work as part of Redcurrant
-Copyright (C) 2015 OpenIO, modified as part of OpenIO Software Defined Storage
+Copyright (C) 2015-2017 OpenIO, as part of OpenIO Software Defined Storage
 
 This library is free software; you can redistribute it and/or
 modify it under the terms of the GNU Lesser General Public
@@ -20,12 +19,12 @@ License along with this library.
 #ifndef OIO_SDS__core__core_h
 # define OIO_SDS__core__core_h 1
 
-# include "core/oiocfg.h"
-# include "core/oioext.h"
 # include "core/oiolog.h"
+# include "core/oioext.h"
 # include "core/oiostr.h"
 # include "core/oiourl.h"
 # include "core/oiocs.h"
 # include "core/oiodir.h"
+# include "core/oiocfg.h"
 # include "core/oiovar.h"
 #endif /*OIO_SDS__core__core_h*/
